mangle the generated pgNN_bindings.rs files into common.rs and pgNN_specific.rs files, where common.rs contains all the top-level items that are exactly the same (as previously rendered by bindgen).  Then the pgNN_specific.rs files contain the items that are specific to them.
This essentially forces development of this library to deal with PG version API differences at the time of writing code.

The bindings for each specific version are still feature-gated, but if the thing we want to use isn't in the common set, then we're forced to deal with it across all the versions in whatever way makes sense for that thing.