refactor: build.rs - avoid io and lexing (#51)
This patch started off as an effort to avoid parsing
the same things multiple times, but that turned out to
be a little tricky. Instead we settle for elimitating
the IO associated with the multiple parses by keeping
the intermediary code in memory as `TokenStream2`s.
This also eliminates lexing a second time (though that's
not likely very useful because lexing is usually an order
of magnitude faster than parsing).

The other point of this patch is the creation of the
new `rewrite_items` routine. Right now this routine is
a pretty useless abstraction, but going forward it should
serve as a good place to keep a pipeline of transformations.
I'm interested in this so that I can inject `Display` instances
for all the types which subclass (according to Postgres' internal
object system) `Node`.