pgx-pg-sys: build.rs - generate Display for nodes
This patch teaches the build.rs script to generate Display
impls for all the generated types which "decend from" postgres
Node types.

For now, we just directly emit a Display instance, but we may
want to instead have a trait that all Nodes implement and then
use blanket impls off of that.

This patch introduces a new StructGraph type which should lay
the groundwork for additional transformations going forward.