Fix issues with error handling across FFI boundaries
This also makes a few functions unsafe, so we had to litter parts of the code with unsafe{} blocks.

The primary user-facing change here is if users are using `pgx::direct_function_call()` to call Rust-based functions with the `#[pg_extern]` annotation, those functions now need to be called using `pgx::direct_pg_extern_function_call()`.

In general, this is yet another follow-up to issue #241.