Add `cargo pgx init --no-restart`
When installing a pgx extension on an already running system, it can be
useful to point `cargo pgx` at that system's `pg_config` in order to
ensure that the extension is being compiled against the correct version.
Right now `cargo pgx init --pgXX` restarts the database, so the only way
to do this is by manually writing the `config.toml`. This commit adds a
flag to `cargo pgx init` `--no-restart` that suppresses the default
behavior, and allows `cargo pgx` to build against an already running
system.