Correctly transform control file version number in sql entity graph
Commit f00c335a7 added the ability to use '@CARGO_VERSION@' as a
placeholder for 'default_version' in the extension's .control file.

That change missed the fact that the contents of the .control file are
loaded into the root node of the sql entity graph. This means that when
using the placeholder, the default_version in the sql entity graph's
control file is '@CARGO_VERSION@'.

This change compiles the package version into the sql-generator, which
performs the replacement at runtime.