feat: extend sql generation via #[pgx(sql)] (#410)
This commit extends the `#[pgx]` and `#[pg_extern]` attributes with a new
`sql` argument type that allows for customizing the behavior of SQL generation
on certain entities in three different ways:

1. `#[pgx(sql = false)]` disables generation of the decorated item's SQL
2. `#[pgx(sql = path::to::function)]` delegates responsibility for
   generating SQL to the named function
3. `#[pgx(sql = "RAW SQL CODE;")]` uses the provided SQL string in place
   of the SQL that would have been generated by default

In the 2nd case, the function has almost the same signature as the
`to_sql` function of the built-in `ToSql` trait, except the function
receives a reference to a `SqlGraphEntity` in place of `&self`. This
allows for extending any of the `SqlGraphEntity` variants using a single
function, as trying to use specific typed functions for each entity type
was deemed overly complex. This also works well with the way `ToSql` is
invoked today, which starts by calling the implementation on a value of
type `SqlGraphEntity`, so we can perform all the checks in a single
place.

As an aside, these custom callbacks can still delegate to the built-in
`ToSql` trait if desired. For example, if you wish to write the
generated SQL for specific entities to a different file.

The motivation for this is that in some edge cases, it is desirable to
elide or modify the SQL being generated. In our case specifically, we
need to chop up the generated SQL so that we can split out non-idempotent
operations separately from idempotent operations in order to properly
manage extension upgrades. Rather than lose the facilities pgx provides,
the `#[pgx(sql)]` attribute allows us to make ad-hoc adjustments to what,
when and where things get generated without needing any upstream support
for those details.