Optionally emit versioned .so (#546)
`pgx` now experimentally supports the option to produce a versioned
shared library. This option is activated by removing the
`module_pathname` parameter in the extension's `.control` file.

When activated, it does the following:

- appends current package version to shared library name
- points generated function SQL to the versioned shared library (instead
  of `MODULE_PATHNAME`)
- replaces `@MODULE_PATHNAME@` placeholder with the path to the versioned
  shared library

Background:

There is an implicit requirement that C extensions maintain ABI
compatibility between versions. Emitting a versioned .so allows for ABI
breaks between versions.

This functionality is experimental and that care must be taken in the
following scenarios:

- when using shared memory
- when using query planner hooks
- when producing SQL schema migrations