Cache Layout in Arrays
This introduces pgx::Layout (with Size and Align types composing it),
parallel to std::alloc::Layout, but modified to allow describing
the Postgres layouts instead.

Some of the constants in this implementation are recently #defined.
We can simply polyfill simple constants and structs in Postgres 10-12,
to make binding against Postgres slightly easier in pgx itself.
This should be done more cautiously for functions and the like.