Make `default!` work with `use pgx::prelude::*`
Previously, PGX has relied on `use pgx::*;` in order to make the framework happen.
However, this severely pollutes the namespaces of our users. We're hogging idents.
Worse, a user collision with our identifiers may result in displacing our logic,
resulting in undesirable and unexpected side effects.

PGX depends on the macros being at least mostly correct in order to be sound,
so this is not at all acceptable.