Make wrappers obey `#[deny(unsafe_op_in_unsafe_fn)]`
The pgx-generated wrapper fn are ubiquitous in pgx expansions,
and they wind up calling the functions of the extension, which
may or may not be unsafe themselves to call.

So, warning or denying the "unmarked unsafe operation in unsafe fn"
lint from Rust RFC 2585 won't get useful data if unsafe fn are called
unless the macro unsafe-wraps its own guarded calls!

Unfortunately in the safe fn call case, this can trigger `unused_unsafe`
lints instead. So, `#[allow(unused_unsafe)]` on the innermost items.