`#[warn(unsafe_op_in_unsafe_fn)]` on srf wrappers
As a more complex function call interface, set-returning fn have
many unambiguously unsafe operations. So, don't unsafe-wrap srf:
Instead set the lint to warn on entire srfs and carefully code so
that unsafe blocks are placed at the right locations to not nest
and not be used on maybe-safe operations.

This can be upgraded to a deny lint later.