Allow denying unsafe ops in PGX-using crates (#707)
* Deny unsafe ops in pg-sys guard code

* Make wrappers obey `#[deny(unsafe_op_in_unsafe_fn)]`

The pgx-generated wrapper fn are ubiquitous in pgx expansions,
and they wind up calling the functions of the extension, which
may or may not be unsafe themselves to call.

So, warning or denying the "unmarked unsafe operation in unsafe fn"
lint from Rust RFC 2585 won't get useful data if unsafe fn are called
unless the macro unsafe-wraps its own guarded calls!

Unfortunately in the safe fn call case, this can trigger `unused_unsafe`
lints instead. So, `#[allow(unused_unsafe)]` on the innermost items.

* `#[warn(unsafe_op_in_unsafe_fn)]` on srf wrappers

As a more complex function call interface, set-returning fn have
many unambiguously unsafe operations. So, don't unsafe-wrap srf:
Instead set the lint to warn on entire srfs and carefully code so
that unsafe blocks are placed at the right locations to not nest
and not be used on maybe-safe operations.

This can be upgraded to a deny lint later.

* Joke slightly less about `unsafe { }`
Avoid `// SAFETY:` annotations that are truly contentless:
They cause us to pass clippy lints that we should not pass.