Allow type structs to be filtered before serialization
In (very) rare circumstances, structs that can be successfully CBOR serialized shouldn't be serialized with everything that was sent in the query, but should instead be stored with only a subset of that data.
Instead of having to write an entire custom inoutfuncs set, this patch allows a type to declare that it needs to "filter" the struct before it's turned into a `Datum`, and then implement a trait to do so.