`#[derive(Deserialize)]` Timestamps
PGX should offer proper deserialization from strings at some point,
however, this doesn't need to be more complicated because
the only user is deserializing from integer data at the moment,
and PGX did not previously directly `impl Deserialize`,
nor did PGX enable the serde feature of the time crate.

It is mildly unfortunate that this means serde doesn't round trip,
currently, but it is preferable to have a working type over not.
This is necessary as previously, PGX's type mappings were
more compatible with the `pg_sys::Timestamp{,Tz}` types.