Problem: working with independent extensions simultaneously
As I am developing a few small extensions, I tend to switch between
them, leaving the database in a state that's not necessarily stable, and
therefore sharing the same instance of PostgreSQL is a bit problematic.

I can use PGX_HOME to use independent installations of PostgreSQL but
they compete for the same set of ports.

Solution: make ports configurable during `pgx init` and in `config.toml`