Golf Display for Node types (#748)
* Golf Display for Node types

This may not seem like much but it removes ~50 KiB from each build.
It also uses more conservative type casts in every case, as casting
`&self as const _ as *mut T` is a bit dubious. It has little effect
in the Rust abstract machine, but it's the wrong signal when
the decl is actually `extern char *nodeToString(const void *obj);`

This thoroughly refactors the safety boundaries so we can be sure
the function is safe, despite handling raw pointers in the impl:
encapsulating the `unsafe fn` entirely inside a private function.