Implement dynamic background worker registration (#710)
* Implement dynamic background worker registration

* Test dynamic background worker execution.

This also changes termination API for background workers a little bit,
enforcing shutdown-scoped limitations on the functionality available.

* Problem: dynamic bgworker shmem initialization

It's not really doing anything as it hooks into postgres initialization

Solution: remove it and put a note in comments

* Improve documentation for background worker handle

* Problem: dynamic bgworker wait_for_* block

This is rather unpleasant behaviour as it is unrecoverable.

Solution: check if the PID was correctly configured and return an error
if it wasn't

* Improve BackgroundWorkerBuilder documentation

It now properly reflects the fact that dynamic background workers are
supported.

* Problem: BackgroundWorkerStatus::Untracked is not self-explanatory

Solution: document it

* Problem: access to `pg_sys::BackgroundWorker` from the builder

In some cases, it is not desirable for the builder to load the
background worker. For example, when actual loading is handled by some
external code that uses this structure from the C API.

Solution: make `BackgroundWorkerBuilder::pg_sys_worker` public

* Problem: BackgroundWorkerBuilder::pg_sys_worker method

Its naming and scoping is problematic, it's not obvious what it does.

Solution: convert it into an `Into` trait implementation

* Problem: imperfect naming for dynamic background worker

From @workingjubilee:

When a handle is a noun, typical Rust form is to use its name unadorned
(almost all our "smart pointers" are also "handles", technically
 speaking). Exceptions are for verbs (JoinHandle) and truly ambiguous
types (e.g. RawHandle and OwnedHandle).

Solution: rename relevant types

* Problem: TerminatingDynamicBackgroundWorker::wait_for_shutdown signature

It doesn't match the signature of
`DynamicBackgroundWorker::wait_for_shutdown`

Solution: ensure they are the same and there's no confusion