Fix issue #792
The crux of the issue here is that `pgx_sys::Datum::from_datum(d, is_null)` can't ask the input Datum if it thinks it's null (via Datum::is_null(&self)), as nullness isn't a concept Datums encode.

A Datum is technically only ever considered null if the accompanying `is_null` argument tells us so.

It so happens that if the Datum type is pass-by-reference, and its value is zero, then it's pointing to a NULL, but Postgres doesn't even guarantee us that that is the nullness indicator.  Again, only the `is_null` argument does that.

As such, when converting a Datum into an Option<Datum>, only listen to the arguments and don't try to infer anything from the input datum's underlying value.

This PR includes a few simple tests that are now green with the change to `from.rs`.

Thanks to @sumerman for catching this!