Problem: PostgresType can't be derived on an enum
In some cases, a Rust enum is not meant to be Postgres enumerated type
but just a type that can be managed and serialized using Rust and its
ecosystem.

A workaround for this is to wrap an enum into a struct and derive
`PostgresType` on it, but it seems like a lot of ceremony for no
benefit.

Solution: allow PostgresType to be derived on an enum

Again, this is not to be confused with `PostgresEnum` which derives
Postgres enumerated type. This one does not.

Closes #726