Introduces a new `PgTryBuilder` type for properly mimicking Postgres' `PG_TRY/PG_CATCH/PG_FINALLY` C macro construct.  This completely replaces the old `pg_try()` function (which is now gone and was severely bugged).
`PgTryBuilder` lets users execute code and respond to any `PgSqlErrorCode` (which now includes a `::RustPanic` variant!) to either do cleanup work and rethrow the error, or otherwise ignore it and return a different value.  Documentation, examples, and tests for PgTryBuilder will be coming in another commit.  Rust drop semantics work properly as well.

There's a lot of other nice changes in here as well...

- The `CONTEXT` ("errcontext_msg") slot of a raised Postgres error now contains as much stack information we can gather while pgx is propagating errors.  Additionally, the `LOCATION` slot contains the `file:line` of whatever raised the first error -- either a Rust panic or a Postgres error.  This did require pushing down a `#[track_caller]` attribute on **every** Postgres FFI we wrap.

Overall, error handling has been cleaned up such that the **only** place we call back to postgres to raise an ERROR (or rethrow one) is the top-level `guard(|| ...)` function, which is used by `#[pg_guard]` and friends.

(note to self: other users of that function throughout the code need to be audited.  I think they need to instead use PgTryBuilder).

- The `ereport!()` macro has been expanded to support a shorthand syntax where the first argument can be the error level name.  Example:  `ereport!(ERROR, PgSqlErrorCode::ERRCODE_INTERNAL_ERROR, "oh no!");"`.  And the `ERROR/PANIC/FATAL` versions know they don't return and insert an `unreachable!()`.

- The `error!()` macro now also supprots specifying an error code:  Example: `error!(PgSqlErrorCode::ERRCODE_NUMERIC_VALUE_OUT_OF_RANGE, "too small")`