cleanup to error handling
- change some variable names
- `PgTryBuilder::finally()` now also runs after a catch handler panics
- `PgTryBuilder` now has a new function to specifically catch Rust panics (`::catch_rust_panic()`)

There was an attempt made to try and track a call stack for better error context message reporting through postgres, but that attempt failed.  I suspect the only nice answer is to use the new Rust `Backtrace` facilities -- at some later date in the future if users complain.

As things stand now, the file:line:col that caused a Rust panic is reported to Postgres to properly fill out the `LOCATION` error slot, and that's good enough.

We also go ahead and copy in a few part of Postgres' `ErrorData` struct when we catch a Postgres error (via our sigsetjmp handling), which may be beneficial to us down the road.  We're not really using it for anything right now.