Problem: Spi::connect is rather limiting
It restricts the signature of the return type. Furthermore, it makes it
much more difficult to re-use an existing connection safely as the only
way to do so is to create an instance of `SpiClient` manually. However,
once we're out of the closure, the connection will be dropped, and code
that relies on that other instance of `SpiClient` will fail.

Solution: introduce reference-counted client instances

Without breaking _documented_ API, this change adds `Spi::client()`
function that returns a reference-counted instance of `SpiClient` that
will finalize the connection when there are no more clones of it left,
making it much easier to have greater control over the control flow.

It _does_ break the API because one can no longer instantiate
`SpiClient` themselves (without using `Spi` API surface). However, this
is an error-prone technique that should rather be avoided.