Problem: SPI API panics on most errors
This is not great as it doesn't allow for better handling of errors.

Solution: redesign SPI signatures to accommodate for errors

Importantly, this now enforces compatibility checks on Datum. This alone
has led me to discovering a few bugs in pgx that are now fixed
(primarily incorrect compatibility checks and wrong expectations in
tests).