Introduce a "codegen" module (publicly exported as `pgx::prelude::__pgx__codegen__private__` that exports all the various pgx and pgx-related symbols that our code generation (via `pgx-macros` and `pgx-sql-entity-graph`) use.
Going forward, if we modify any code generation or support code generation for some new construct, we will need to make sure any new symbols are added to the `codegen.rs` module and referenced as `__pgx__codegen__private__::SymbolName`.

The primary justification for this is that it'll make it easier for certain types of extensions (read: plrust) to limit the scope of pgx internals they need to publicly expose.  And while it'll be a bit of a maintenance burden for us, it is very nice to have a single module that lists all the symbols our code generation uses -- this will be handy for auditing purposes among others, I'm sure.

Also touch `pgx::prelude` a bit, adding a few more things to help existing code compile.  It also updates the tests (doctest and unit tests) along with examples that needed it to `use pgx::prelude::*`.

I'd expect this commit to be a bit of a breaking change for existing users where perhaps they're not doing `use pgx::prelude::*` in their code.

I have a thought of a follow-up commit to *reduce* the rexports from the top-level `pgx::` module, but we can save that for another day.