Problem: dropping owned memory context while it's current
The error will look like this:

```
TRAP: FailedAssertion("context != CurrentMemoryContext", File: "mcxt.c", Line: 224, PID: 71536)
```

Solution: avoid failing this assertion

We can say that it is on the user of `pgx` to ensure that the owned
context is dropped at the right time, but it may get tricky to do so.

For example, certain types are not UnwindSafe, so we can't use
`switch_to` or `PgTryBuilder` and guard the exceptional situation to
ensure the context is no longer current before it is dropped. So, due
to these limitations, we're forced to use `set_as_current` and the
reversal to the old context will never happen if there's a panic.