Problem: switching to parent on owned memory context's drop
If owned memory context is current, it is being switched to its parent
prior to deletion as we'll otherwise have a failed assertion from within
Postgres. This is introduced in #921.

However, this is still not perfect as it may leave us in an unexpected
context (parent) as opposed to the previous one.

Solution: remember the previous memory context for owned contexts
and restore it on drop if necessary.

This does introduce a small breaking change to the API as it requires
`set_as_current` to take a mutable reference to `Self`