Ensuring owned memory context is not current when dropping it (extension) (#922)
The error will look like this:

```
TRAP: FailedAssertion("context != CurrentMemoryContext", File: "mcxt.c", Line: 224, PID: 71536)
```

Solution: avoid failing this assertion

We can say that it is on the user of `pgx` to ensure that the owned
context is dropped at the right time, but it may get tricky to do so.

For example, certain types are not UnwindSafe, so we can't use
`switch_to` or `PgTryBuilder` and guard the exceptional situation to
ensure the context is no longer current before it is dropped. So, due
to these limitations, we're forced to use `set_as_current` and the
reversal to the old context will never happen if there's a panic.

If owned memory context is current, it is being switched to its parent
prior to deletion as we'll otherwise have a failed assertion from within
Postgres. This is introduced in #921.

However, this is still not perfect as it may leave us in an unexpected
context (parent) as opposed to the previous one.

Solution: remember the previous memory context for owned contexts
and restore it on drop if necessary.

This does introduce a small breaking change to the API as it requires
`set_as_current` to take a mutable reference to `Self`