Problem: setting owned memory context as current twice
When an owned context is set as current twice, dropping it will fail or
result in current context being deleted.

Solution: ensure we're not forgetting the "previous" context

When we set an owned context as a current one, but it is already
current, it would previously overwrite the reference to the "previous"
context to point to itself. At which point we don't know the previous
context we can revert to anymore.

This change ensures setting an owned context as a current one in an
uninterrupted sequence is idempotent.