`PgMemoryContexts::get_context_for_pointer()` is wildly unsafe. (#947)
* `PgMemoryContexts::get_context_for_pointer()`` is wildly unsafe.

So much so that it's now been removed (it was a dumb wrapper around `pgx_GetMemoryContextChunk` anyways in favor of a type that forces the user to go through an `unsafe{}` block to get a `PgMemoryContexts::Of()` variant.

Also, the `pgx_GetMemoryContextChunk` function has been ported to Rust (removed from the c-shim) and moved over to pgx-pg-sys as `GetMemoryContextChunk`.  And with it came a new `MemoryContextIdValid` function.

Note that this also required ripping off the `TYPEALIGN` and `MAXALIGN` functions from PR #948, so once it and this is merged, there needs to be some kind of follow-up.