Port `c-shim.c`'s `pgx_ereport()` function to rust (#946)
* Port `c-shim.c`'s `pgx_ereport()` function to rust

This rewrites the C version of `pgx_ereport()` as Rust, as `pgx-pg-sys/src/submodules/panic.rs:do_ereport()`.  Parts of it are under Postgres version feature flags, because of course Postgres' error reporting system is different.

It also blocks a small list of "err*" functions from our bindings (in `build.rs`) and also adds support for the "hint" error line, which has a default of None.