Teach `pgx-tests/src/framework.rs` to detect `cargo test` feature arguments (#967)
Running a plain `cargo test --features "a b c"` never properly worked with pgx' test framework as none of the "feature"-related arguments were being passed down into pgx' test framework.  Now they are.

In the case of a `cargo test --all --features "a b c"`-style execution, `cargo-pgx` will elide passing features to a workspace member that doesn't have that feature.

This necessitated reworking how **all** of the `cargo-pgx` commands detect which Postgres version feature flag to use, along with the general set of features to turn on.  The search order is now:

### schema/run/test/connect

- direct `cargo-pgx` argument
- `--features "pgXX"` flag
- `default = [ "pgXX" ]` from the extension's Cargo.toml

### install/package

- created from the `pg_config` on the system `$PATH` or the `--pg_config` argument

### start/stop/status

- direct `cargo-pgx` argument
- `default = [ "pgXX" ]` from the extension's Cargo.toml

As a few drive-bys, teach cargo-pgx to output the type of `pgXX` feature flag it found and the full path to the `pg_config` binary it's using for that operation.