Add a default `cshim` feature flag (#958)
This adds a `cshim` feature flag to `pgx`, enabled by default, and pushes it down to `pgx-pg-sys` and `pgx-tests`.

With this turned off we lose low-level access to a few things:

   - `pgx_list*` functions
   - `pgx_planner_rt_fetch()`
   - `pgx_SpinLock*` functions

From `pgx`' perspective, this means a few modules are just not available without `cshim`:

   - `hooks` (uses `PgList`)
   - `list` (impl for `PgList` which uses `pgx_list*` functions)
   - `namespace` (uses `PgList`)
   - `spinlock` (uses `pgx_SpinLock*` functions)

This feature flag will hopefully make it easier to cross-compile pgx with only a small bit of compromises.  This feature flag was initially designed with [plrust](https://github.com/tcdi/plrust) user functions in mind, and they won't need access to any of the above.