Change bounds on `IntoDatum for Result<T, E>` such that `E: Any + Display`
Broaden the bounds on our support for converting a Result into a datum a bit so that we can detect if the `Err` variant contains one of our `ErrorReport` types.  If it does, that's the thing we'll raise, at the `PgLogLevels::ERROR` level.  Otherwise we'll just raise a `panic!()` with the to_string() of the error as before.

This enables users to return `Result::Err` variants that contain a more rich set of error reporting information such as a specific `PgSqlErrorCode` and optionally "hint" and "detail" lines.