Support returning `Result<TableIterator/SetOfIterator, E>`
Add the ability to return a Result of a TableIterator or SetOfIterator.  This just helps to make things more consistent with the other cases where we can now return a Result of something.

Also introduce a new trait, `pg_sys::panic::ErrorReportable` that when its `.report()` method is called will return a value or raise an error depending on the context.  We provide a default implementation for `Result<T, E: Any+Display>` that understands how to downcast the error to look for an `ErrorReport` instance.  This moves that code out of the `impl IntoDatum for Result` code.