Fix a bug trying to return `Result<pg_sys::Oid>`
HACK for `Result<T, E>` -- we replace the `E` with an underscore

specifically, this is the cases where we need to use the `map_source_only!()` macro to
create a Rust type to SQL type mapping.  As of this writing that's only for the `pg_sys::Oid`
type, but we'll do our best to generalize this.

What we want to do is rewrite any kind of `Result<T,E>` pattern into `Result<T,_>`, which
is how the `map_source_only!()` macro does it.