Set returning `#[pg_extern]` functions can cause segfault
If a `#[pg_extern]`-style function raises a Postgres ERROR during its iteration process, Postgres will segfault.

This fixes the problem by not trying to re-create/re-leak the pointer we initially leaked with `PgMemoryContexts.leak_and_drop_on_delete/leak_trivial_alloc()`.  Instead, we track it in our "iterator holder" as a `NonNull` pointer, which is all we need.