Move generated SRF code into managed code (#987)
This takes out nearly all of the code generation for `TableIterator`- and `SetOfIterator`-returning functions and moves it into actual code in the new `pgx/src/srf.rs`.
    
Also updates the various `srf_xxx()` functions in `pgx/src/fcinfo.rs` to not use PgBox -- that was showing up in profiling, even in --release.
    
We also have a new trait called `IntoHeapTuple` that knows how to convert, given a TupleDesc, Rust tuples (up to 32 elements) into a Postgres `pg_sys::HeapTuple`.  It requires that each element be `IntoDatum`.  This is used by the srf implementation for `TableIterator`.