Extend Spi's understanding of "readonly" to transaction
Mucking with the pushing a new Snapshot if we *think* we've modified the database isn't correct.  It can cause problems elsewhere within Postgres where Postgres expects the Portal's snapshot to be the same as when the Portal was created.

The best I can see to do is just track "readonly" (renamed to `MUTABLE_MODE`) for the duration of the current transaction.