Cleanup `PgTrigger` such that its public API can now be safe.
This pushes more responsibility on to the user constructing a `PgTrigger`, but that responsibility was already there anyways.

We also add the concept of an `Old` trigger tuple, to go along with the existing `New` and `Current`, and properly implement creating these things.

The `PgTriggerSafe` struct is completely deleted -- it's no longer necessary.