`PgTrigger` now has a lifetime and borrows from its source `pg_sys::FunctionCallInfoBaseData`.
This ensures that a `PgTrigger` instance, which essentially borrows pointers from an "fcinfo" doesn't outlive the "fcinfo" from which it was created.  We also clean up the members of `PgTrigger` to just the bare minimum we need.

Another benefit here is that the `PgTrigger::relation(&self)` function can now directly use the `RelationData` pointer that's already been opened for us by Postgres.

Adding a lifetime to `PgTrigger` caused a bunch of touch points in our tests and examples, so I went ahead and fixed another issue at the same time.  `#[pg_trigger]` functions now must return `Result<Option<PgHeapTuple>, E>`.  It is desired that a trigger function be able to "return NULL" to indicate that the operation that caused it to fire should **not** be performed.  Prior to this change there was no way to do that.  A test has been added for this case as well.