Memoize DB encoding for faster validation
This performs a check for whether the database is UTF-8,
and then uses that information to reduce text validation.
An initial check is performed at database load time.
If the result discovers a completely incompatible,
not-even-ASCII encoding, then the extension will bail.
This shouldn't happen as they aren't supported server-side
but maybe Postgres will implement support in the future?