Support building against macOS universal binaries
FIXES: #1087

A "universal binary" is a fat binary which contains multiple builds for
different architectures. When Postgres is packaged in this format it's
necessary to slice out the build for the current architecture before
trying to parse it. The program terminates early if the universal binary
doesn't support the current architecture.

The list of supported architectures in `slice_arch32` is limited by
[`object`][object-arch], which supports fewer than [Rust][rust-arch].

The universal binary test fixture was simply chosen because of it's
small size relative to other binaries in Postgres.

[^1]: https://github.com/gimli-rs/object/issues/454#issuecomment-1200668989

[rust-arch]: http://doc.rust-lang.org/1.68.2/std/env/consts/constant.ARCH.html
[object-arch]: https://docs.rs/object/latest/object/read/macho/trait.FatArch.html#method.architecture