Add Array slice support (#1147)
Add support for getting a slice from an `Array<T>` where `T` is a one of these primitives:  i8, i16, i32, i64, f32, f64.  It's possible that we could support other fixed-size types as well, but this is a good place to start.

We also implement `IntoDatum` for `Array<T>`, add some tests, and expand the `arrays` example a little bit.