Type roundtrip tests (#1185)
A new set of tests to simply check that we can "roundtrip" a Datum.  That is, we return the same value as we were given in a `#[pg_extern]` function argument.

Essentially this tests that the type's `FromDatum` and `IntoDatum` impls match, and in the case of arrays, tests that the iterator impl for `Array<T>` works for its individual datums.

In doing so, this uncovered a bug with converting a `pg_sys::Point` (and `pg_sys::BOX`) into a Datum, so those are fixed in here as well.

The addition of `#[derive(Copy)]` to the "Complex" type in pgrx-tests is just to make it work well with the roundtrip tests here.